break up monolithic trig file for more incremental compiling
