+++ app/tool/trig/algo.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "../pftool.h"
+
+void algo(const std::string& cmd, Target* tgt);
+
