+++ app/tool/trig/align.h
@@ -1,0 +1,3 @@
+#pragma once
+#include "../pftool.h"
+void align(const std::string& cmd, Target* tgt);
