+++ app/tool/trig/histo.h
@@ -1,0 +1,3 @@
+#pragma once
+#include "../pftool.h"
+void histo(const std::string& cmd, Target* tgt);
