+++ app/tool/trig/self_trig.h
@@ -1,0 +1,3 @@
+#pragma once
+#include "../pftool.h"
+void self_trig(Target* tgt);
